+++ CHANGELOG.md
@@ -2,6 +2,7 @@ v2.5.9 (FUTURE)
 
 ## Enhancements
 
+* [#2933](https://github.com/digitalocean/netbox/issues/2933) - Add username to outbound webhook requests
 * [#3011](https://github.com/digitalocean/netbox/issues/3011) - Add SSL support for django-rq (requires django-rq v1.3.1+)
 
 ## Bug Fixes